requirements.txt
# Core ML + Data Science
pandas==1.5.3
numpy==1.23.5
scikit-learn==1.4.2
matplotlib==3.8.4
seaborn==0.12.2
joblib==1.3.2
jupyter==1.0.0

# MLflow for experiment tracking
mlflow==2.11.3

# Airflow (core + tested dependency versions for Python 3.11)
apache-airflow==2.8.1
email-validator==1.3.1
pendulum==2.1.2  # Airflow runtime requirement
flask-session==0.4.0 # Airflow dependency


# API
fastapi==0.103.2
uvicorn==0.23.2
pydantic==2.3.0
pyyaml==6.0.1
